+++ src/index.js
@@ -1,7 +1,7 @@
 // Copyright (c) Cosmo Tech.
 // Licensed under the MIT license.
 
-export { CreateScenarioButton, SelfDestructLinkButton, SignInButton } from './buttons';
+export { CreateScenarioButton, SelfDestructLinkButton, SignInButton, RolesEditionButton } from './buttons';
 export { ScenarioNode } from './cards';
 export { CytoViz, Dashboard, ScenarioManagerTreeList, SimplePowerBIReportEmbed } from './charts';
 export { SimpleTwoActionsDialog, DontAskAgainDialog } from './dialogs';
@@ -17,13 +17,16 @@ export {
   UPLOAD_FILE_STATUS_KEY,
   Table,
   TABLE_DATA_STATUS,
+  RoleEditor,
+  SelectWithAction,
 } from './inputs';
 export { UserInfo, HelpMenu } from './menus';
 export {
+  ErrorBanner,
   ErrorsPanel,
   FixedRatioContainer,
-  ScenarioValidationStatusChip,
   LoadingLine,
-  ErrorBanner,
+  PermissionsGate,
+  ScenarioValidationStatusChip,
   DefaultAvatar,
 } from './misc';