+++ src/inputs/RoleEditor/index.js
@@ -1,0 +1,4 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+export { RoleEditor } from './RoleEditor';