+++ src/inputs/index.js
@@ -6,3 +6,5 @@ export { UploadFile, UPLOAD_FILE_STATUS_KEY } from './UploadFile';
 export { Table, TABLE_DATA_STATUS } from './Table';
 export { BasicEnumInput, BasicTextInput, BasicNumberInput, BasicToggleInput, BasicDateInput } from './BasicInputs';
 export { EditableLabel } from './EditableLabel';
+export { RoleEditor } from './RoleEditor';
+export { SelectWithAction } from './SelectWithAction';